ARM: dts: mvebu: Introduce a new compatible for the PMSU node
The initial binding for PMSU were wrong. It didn't take into account
all the register from the PMSU and in the other hand refered to
registr which are not part of PMSU. This commit introduce a new
compatible for the PMSU node and mark the previous one as
obsolete. However we still continue to support it during a few release.

Cc: devicetree@vger.kernel.org
Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>