ARM: kernel: add special case for SCU base address detection for Marvell SoCs
The Armada 375 SoC uses the Cortex-A9 and therefore provides a
SCU. The Marvell SoCs have a configurable internal register base
address: it is 0xd0000000 by default when the processor boots, but the
bootloaders remap it at 0xf1000000 to leave more space below the 4 GB
limit for normal memory.

Unfortunately, the reading of the CP15 register that gives the SCU
base address doesn't take into account this remapping, and returns
0xd000c000 regardless of whether the internal registers have been
remapped at a different location or not.

This commit adds a fixup to the SCU base address that consists in
detecting if the returned address is 0xd000c000, and by replacing it
by 0xf100c000 in this case. This code is executed really early, and we
unfortunately don't have other really good ways of detecting whether
we're running on a problematic SoC or not.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>