ARM: mvebu: enable the SDHCI interface on Armada 385
In commit "mmc: sdhci-pxav3: add support for the Armada 38x SDHCI
controller", the sdhci-pxav3 driver has been extended to also be
usable on Armada 38x platforms.

Therefore, this commit adds the necessary Device Tree informations to
declare this SDHCI interface in the Armada 38x SoC, and also in the
Armada 385 Development Board.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>