ARM: mm: enable PJ4B SMP bit unconditionally on Armada XP
One of the requirement for I/O coherency to work properly on Armada XP
is to have the SMP bit set in the Auxiliary Functional Modes Control
Register 0, and this regardless of whether we're compiled CONFIG_SMP
or not.

However, while this bit exists in the PJ4B-MP core (used in the Armada
XP), it doesn't exist in the PJ4B core (used in the Armada 370), so we
add a test to verify we're running on a PJ4B-MP core.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>