ARM: mvebu: define the Armada 370/375/38x/XP machine_desc flags
The Armada 370 is a purely non-SMP processor, so it does not need the
MACHINE_NEEDS_SHAREABLE_PAGES flag, and more importantly this flag
cannot be set with this CPU core.

For all other processors, they are SMP-capable, but have cut-down
variants that only have one core. And for these variants, we need both
the cache policy to be write-allocate, and the shareable attribute to
be set on page tables in order for I/O coherency to work properly.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>