ARM: mvebu: Add a common function for the boot address work around
On some of the mvebu SoC and due to internal BootROM issue, the CPU
initial jump code should be placed in the SRAM memory of the SoC. In
order to achieve this, we have to unmap the BootROM and at some
specific location where the BootROM was placed, create a specific MBus
window for the SRAM. This SRAM is initialized with a few instructions
of code that allows to jump to the real secondary CPU boot
address. The SRAM used is the Crypto engine one.

This work around is currently needed for booting SMP on Armada 375 Z1
and will be needed for cpuidle support on Armada 370. Instead of
duplicating the same code, this commit introduces a common function to
handle it: mvebu_setup_boot_addr_wa().

Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>