ARM: mvebu: Add function to export the physical address of the boot register
In order to boot the secondary CPUs on Armada 375 Z1, we need to read
the boot address of these CPUs through a register part of the System
Controller. This done very early where MMU is not enable yet, so we
need to get the physical address of this register to use it.

In the previous version of this workaround the physical address were
hardcoded whereas depending of the Mbus configuration this address may
change.

This commit will allow to use the generic boot workaround function
with the correct address.