ARM: mvebu: Use the common function for Armada 375 SMP workaround
Use the common function mvebu_setup_boot_addr_wa() introduced in the
commit ARM: mvebu: Add a common function for the boot address work
around" "instead of the dedicated version for Armada 375.

This commit also moves the WA in the system-controller module. Indeed
the workaround on 375 is really related to setting the boot address
which is done by the system controller.

In bonus we don't use anymore an harcoded value to access the register
storing the boot address.

Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>