ARM: mvebu: Add workaround for cpuidle support for Armada 370
On Armada 370, there is "a slow exit process from the deep idle state
due to heavy L1/L2 cache cleanup operations performed by the BootROM
software". To avoid this, we replace the restart code of the BootROM
by a simple jump to the boot address. Then the code located at this
boot address will take care of the initialization.

For this purpose, we use the common function mvebu_boot_addr_wa()
introduced in the previous commit.

Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>