ARM: mvebu: use the cpufreq-dt platform_data for independent clocks
This commit adjusts the registration of the cpufreq-dt driver in the
mvebu platform to indicate to the cpufreq driver that the platform has
independent clocks for each CPU.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
---
Jason, this commit was part of my "[PATCHv2 0/4] cpufreq-dt,
platform_data based proposal" series, which included 3 patches for
drivers/cpufreq/, and this last patch for mach-mvebu to make use of
the new cpufreq feature. However, in order to not confuse the cpufreq
maintainer, when I sent "[PATCH v3 0/3] cpufreq: cpufreq-dt:
platform_data based proposal", I didn't include this patch for
mach-mvebu.

But now, it means that this is the only missing piece to get cpufreq
working on Armada XP. Do you think you could squeeze it into the
things merged for 3.19 ? It's a very small patch, so hopefully it's
still possible to do so.

The patch is based on mvebu/soc.