ARM: mvebu: Clean-up the SMP support for the Armada XP
This patch finishes converting the SMP support for the Armada XP to
the cpu method style used with the device tree. It allows removing the
smp_op filed in the dt machine struct.

From this point, it allows moving all the content of the
armada-370-xp.h in their respective files, and at the end removing
this header.

This patches finally removes all the references to the armada-370-xp.h
header.

Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>