usb: ehci-orion: keep phy_version in struct orion_ehci_hcd
Currently, the phy_version is only used in the ->probe() function, and
therefore a local variable is used. However, with the upcoming
introduction of suspend/resume support, we will need to access this
information from the ->resume() function.

In order to achieve that, a new 'phy_version' field is added in
orion_ehci_hcd, and is used in ->probe() in replacement of the local
variable.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>