ARM: mvebu: use IRQ macros to define the SDHCI interrupt on Armada 38x
The Armada 38x SDHCI controller definition in the SoC Device Tree file
is using hardcoded values instead of the nice GIC and IRQ macros
provided by DT header files. This commit fixes that, which allows the
SDHCI interrupt definition to be consistent with was is done in the
rest of the file.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>