arm: kernel: hardcode SCU address
Early versions of Marvell bootloaders do not set the correct SCU
address in the dedicated CP15 register, causing the kernel to fail
booting. This ugly hack hardcodes the correct SCU address, so that the
kernel doesn't pay attention to the one passed by the bootloader.

This ugly hack will never be pushed to the mainline kernel since it's
breaking multiplatform support completely. The real fix is to update
your bootloader.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>